move utility to utils
